# docs/static/swagger.json
{
...
  "info": {
    "description": "Documenta\u00e7\u00e3o autom\u00e1tica das rotas do sistema com arquitetura DDD e SQL Puro.",
    "title": "Documenta\u00e7\u00e3o da API - Projeto g03",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/operacoes/": {
      "get": {
        "parameters": [
          {
            "default": 1,
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "in": "query",
            "name": "per_page",
            "type": "integer"
          },
          {
            "description": "Filtrar por status: ativa, pausada, encerrada",
            "in": "query",
            "name": "status",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de opera\u00e7\u00f5es"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lista todas as opera\u00e7\u00f5es com pagina\u00e7\u00e3o e filtro por status",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "created_at": {
                  "example": "2024-03-15T08:00:00",
                  "type": "string"
                },
                "localizacao": {
                  "example": "Avenida Paulista, SP",
                  "type": "string"
                },
                "name": {
                  "example": "Opera\u00e7\u00e3o Centro",
                  "type": "string"
                },
                "status": {
                  "example": "ativa",
                  "type": "string"
                }
              },
              "required": [
                "name",
                "status",
                "localizacao",
                "created_at"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Opera\u00e7\u00e3o criada com sucesso"
          },
          "400": {
            "description": "Dados inv\u00e1lidos"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Cria uma nova opera\u00e7\u00e3o de campo",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      }
    },
    "/api/operacoes/{oid}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "oid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opera\u00e7\u00e3o removida"
          },
          "404": {
            "description": "Opera\u00e7\u00e3o n\u00e3o encontrada"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Remove uma opera\u00e7\u00e3o (e em cascata todos seus drones e scans)",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "oid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da opera\u00e7\u00e3o"
          },
          "404": {
            "description": "Opera\u00e7\u00e3o n\u00e3o encontrada"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Busca uma opera\u00e7\u00e3o pelo UUID",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "oid",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "localizacao": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "status": {
                  "description": "ativa | pausada | encerrada",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opera\u00e7\u00e3o atualizada"
          },
          "404": {
            "description": "Opera\u00e7\u00e3o n\u00e3o encontrada"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Atualiza nome e/ou status de uma opera\u00e7\u00e3o",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      }
    },
    "/api/operacoes/{oid}/drones": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "oid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de drones da opera\u00e7\u00e3o"
          },
          "404": {
            "description": "Opera\u00e7\u00e3o n\u00e3o encontrada"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lista todos os drones de uma opera\u00e7\u00e3o espec\u00edfica",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      }
    },
    "/api/operacoes/{oid}/status": {
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "oid",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "status": {
                  "example": "encerrada",
                  "type": "string"
                }
              },
              "required": [
                "status"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status atualizado"
          },
          "400": {
            "description": "Status inv\u00e1lido"
          },
          "404": {
            "description": "Opera\u00e7\u00e3o n\u00e3o encontrada"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Muda o status de uma opera\u00e7\u00e3o (atalho r\u00e1pido)",
        "tags": [
          "Opera\u00e7\u00f5es"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Insira o token JWT no formato: Bearer {seu_token}",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}